Move mock data fixtures for Style API to conftest.py
