+++ tests/conftest.py
@@ -1,7 +1,11 @@
 import datetime
 from unittest.mock import MagicMock
 
+import numpy as np
 import pytest
+import xarray as xr
+
+from tests.utils import dummy_da, coords, dim1_da
 
 
 @pytest.fixture
@@ -271,3 +275,125 @@ def wcs_global_cfg():
         },
         "native_format": "GeoTIFF",
     }
+
+
+@pytest.fixture
+def dummy_raw_data():
+    output = xr.Dataset({
+        "ir": dummy_da(3, "ir", coords),
+        "red": dummy_da(5, "red", coords),
+        "green": dummy_da(7, "green", coords),
+        "blue": dummy_da(2, "blue", coords),
+        "uv": dummy_da(-1, "uv", coords),
+    })
+    return output
+
+
+@pytest.fixture
+def null_mask():
+    return dummy_da(True, "mask", coords, dtype=np.bool)
+
+
+@pytest.fixture
+def dummy_raw_calc_data():
+    dim_coords = [-2.0, -1.0, 0.0, -1.0, -2.0, -3.0]
+    output = xr.Dataset({
+        "ir": dim1_da("ir", [800, 100, 1000, 600, 200, 1000], dim_coords),
+        "red": dim1_da("red", [200, 500, 0, 200, 200, 700], dim_coords),
+        "green": dim1_da("green", [100, 500, 0, 400, 300, 200], dim_coords),
+        "blue": dim1_da("blue", [200, 500, 1000, 600, 100, 700], dim_coords),
+        "uv": dim1_da("uv", [400, 600, 900, 200, 400, 100], dim_coords),
+        "pq": dim1_da("pq", [0b000, 0b001, 0b010, 0b011, 0b100, 0b111], dim_coords,
+                      attrs={
+                                "flags_definition": {
+                                    "splodgy": {
+                                        "bits": 2,
+                                        "values": {
+                                            '0': "Splodgeless",
+                                            '1': "Splodgy",
+                                        },
+                                        "description": "All splodgy looking"
+                                    },
+                                    "ugly": {
+                                        "bits": 1,
+                                        "values": {
+                                            '0': False,
+                                            '1': True
+                                        },
+                                        "description": "Real, real ugly",
+                                    },
+                                    "impossible": {
+                                        "bits": 0,
+                                        "values": {
+                                            '0': False,
+                                            '1': "Woah!"
+                                        },
+                                        "description": "Won't happen. Can't happen. Might happen.",
+                                    },
+                                }
+                            })
+    })
+    return output
+
+
+def dim1_null_mask(coords):
+    return dim1_da("mask", [True] * len(coords), coords)
+
+
+@pytest.fixture
+def raw_calc_null_mask():
+    dim_coords = [-2.0, -1.0, 0.0, -1.0, -2.0, -3.0]
+    return dim1_da("mask", [True] * len(dim_coords), dim_coords)
+
+
+@pytest.fixture
+def dummy_col_map_data():
+    dim_coords = [-2.0, -1.0, 0.0, -1.0, -2.0, -3.0]
+    output = xr.Dataset({
+        "pq": dim1_da("pq", [0b01000, 0b11001, 0b01010, 0b10011, 0b00100, 0b10111], dim_coords,
+                      attrs={
+                          "flags_definition": {
+                              "joviality": {
+                                  "bits": 3,
+                                  "values": {
+                                      '0': "Melancholic",
+                                      '1': "Joyous",
+                                  },
+                                  "description": "All splodgy looking"
+                              },
+                              "flavour": {
+                                  "bits": 3,
+                                  "values": {
+                                      '0': "Bland",
+                                      '1': "Tasty",
+                                  },
+                                  "description": "All splodgy looking"
+                              },
+                              "splodgy": {
+                                  "bits": 2,
+                                  "values": {
+                                      '0': "Splodgeless",
+                                      '1': "Splodgy",
+                                  },
+                                  "description": "All splodgy looking"
+                              },
+                              "ugly": {
+                                  "bits": 1,
+                                  "values": {
+                                      '0': False,
+                                      '1': True
+                                  },
+                                  "description": "Real, real ugly",
+                              },
+                              "impossible": {
+                                  "bits": 0,
+                                  "values": {
+                                      '0': False,
+                                      '1': "Woah!"
+                                  },
+                                  "description": "Won't happen. Can't happen. Might happen.",
+                              },
+                          }
+                      })
+    })
+    return output
